requirements.txt
fastapi
uvicorn
websockets
langchain
langgraph
openai
requests
pyyaml
pinecone
dotenv
murf
IPython